Moved 4 Functions to functions.js and Moved Preconnect to Upper Section
