+++ public/index.html
@@ -6,12 +6,12 @@
      <meta name="author" content="ArsenTech">
      <meta name="viewport" content="width=device-width, initial-scale=1.0">
      <meta name="theme-color" content="#22b455">
-     <link rel="stylesheet" media="screen" href="css/style.css">
-     <link rel="icon" href="files/favicon.ico">
-     <link rel="apple-touch-icon" href="files/apple-touch.png">
      <link rel="dns-prefetch" href="https://fonts.gstatic.com">
      <link rel="dns-prefetch" href="https://firebasestorage.googleapis.com/">
      <link rel="dns-prefetch" href="https://fonts.googleapis.com">
+     <link rel="stylesheet" media="screen" href="css/style.css">
+     <link rel="icon" href="files/favicon.ico">
+     <link rel="apple-touch-icon" href="files/apple-touch.png">
      <title>Downloads | ArsenTech's Website</title>
 </head>
 <body>
@@ -64,7 +64,7 @@
                          <li class="list active" data-filter="all">All</li>
                          <li class="list" data-filter="windows">Windows Versions</li>
                          <li class="list" data-filter="softwares">Software</li>
-                         <li class="list" data-filter="internet-explorer">Internet Explorer Versions</li>
+                         <li class="list" data-filter="internet-explorer">Internet Explorer</li>
                          <li class="list" data-filter="ms-dos">MS-DOS Versions</li>
                          <li class="list" data-filter="updates">Updates</li>
                     </ul>
