+++ public/js/functions.js
@@ -1,3 +1,5 @@
+import {ref, getDownloadURL, getMetadata } from "https://www.gstatic.com/firebasejs/9.13.0/firebase-storage.js";
+import {storage} from "./firebase.js"
 function displayCards(cards, cardList){
      cardList.innerHTML = cards.map((card, index)=>{
           return `
@@ -22,22 +24,26 @@ function formatBytes(t,B=2){
      a=Math.floor(Math.log(t)/Math.log(1024));
      return`${parseFloat((t/Math.pow(1024,a)).toFixed(o))} ${["Bytes","KB","MB","GB","TB","PB","EB","ZB","YB"][a]}`
 }
-function lazyCss(e) {
-     const t = document.createElement("link");
-     t.href = e, t.rel = "stylesheet"; 
-     t.type = "text/css";
-     document.getElementsByTagName("head")[0].appendChild(t);
-};
-function lazyJS(e){
-     const t = document.createElement("script");
-     t.src = e;
-     t.defer = true; 
-     t.type="module"; 
-     document.body.appendChild(t)
+function lazyCss(e) {const t = document.createElement("link");t.href = e, t.rel = "stylesheet"; t.type = "text/css";document.getElementsByTagName("head")[0].appendChild(t);};
+function lazyJS(e){const t = document.createElement("script");t.src = e;t.defer = true; t.type="module"; document.body.appendChild(t)}
+function downloadFile(file){getDownloadURL(ref(storage, file)).then(url => {const link = document.createElement("a");link.href = url;link.download = "";link.click();}).catch((error) => console.error(error));}
+const downloadWallpaper = (file) => getDownloadURL(ref(storage, file)).then(url => window.open(url)).catch((error) => console.error(error));
+function getFileInfo(file, index){
+     const fileInfoRef = ref(storage, file);
+     getMetadata(fileInfoRef).then(data => {
+          const sizeDisplay = document.querySelectorAll("#size"),
+          dateDisplay = document.querySelectorAll("#updated");
+          sizeDisplay[index].innerHTML = formatBytes(data.size);
+          dateDisplay[index].innerHTML = data.updated.slice(0,10);
+     }).catch((error) => console.error(error));
 }
+function getWallpaperInfo(f,i){const fileInfoRef = ref(storage, f);getMetadata(fileInfoRef).then(data => {const wallpaperSizeDisplay = document.querySelectorAll("#fileSize");wallpaperSizeDisplay[i].innerHTML = formatBytes(data.size);}).catch((error) => console.error(error));}
 export {
      displayCards,
-     formatBytes,
      lazyCss,
-     lazyJS
+     lazyJS,
+     downloadFile,
+     downloadWallpaper,
+     getFileInfo,
+     getWallpaperInfo,
 }
