+++ public/js/script.js
@@ -1,7 +1,5 @@
-import {ref, getDownloadURL, getMetadata } from "https://www.gstatic.com/firebasejs/9.13.0/firebase-storage.js";
-import {storage} from "./firebase.js"
 import { downloads } from "./data.js";
-import { formatBytes, displayCards, lazyJS, lazyCss } from "./functions.js";
+import { displayCards, lazyJS, lazyCss, downloadFile, downloadWallpaper, getFileInfo, getWallpaperInfo } from "./functions.js";
 const navbar = document.querySelector( ".navbar" ),
       toggler = document.querySelector( ".menu-toggler" ),
       navMenu = document.querySelector( ".navbar-menu" ),
@@ -28,27 +26,10 @@ document.getElementById("yearCount").innerHTML=(new Date).getFullYear();
 function toggleActive(){toggler.classList.toggle("active"); navMenu.classList.toggle("active");}
 function toggleMode(){
      document.body.classList.toggle("dark");
-     if(!document.body.classList.contains("dark")){
-          modeToggler.querySelector("img").src = "files/light.svg";
-          localStorage.setItem("arsentech-theme", "light");
-     } else {
-          modeToggler.querySelector("img").src = "files/dark.svg";
-          localStorage.setItem("arsentech-theme", "dark");
-     }
+     if(!document.body.classList.contains("dark")){modeToggler.querySelector("img").src = "files/light.svg";localStorage.setItem("arsentech-theme", "light");} 
+     else {modeToggler.querySelector("img").src = "files/dark.svg";localStorage.setItem("arsentech-theme", "dark");}
 }
 function closeMenu(){toggler.classList.remove("active"); navMenu.classList.remove("active");}
-function downloadFile(file){getDownloadURL(ref(storage, file)).then(url => {const link = document.createElement("a");link.href = url;link.download = "";link.click();}).catch((error) => console.error(error));}
-const downloadWallpaper = (file) => getDownloadURL(ref(storage, file)).then(url => window.open(url)).catch((error) => console.error(error));
-function getFileInfo(file, index){
-     const fileInfoRef = ref(storage, file);
-     getMetadata(fileInfoRef).then(data => {
-          const sizeDisplay = document.querySelectorAll("#size"),
-          dateDisplay = document.querySelectorAll("#updated");
-          sizeDisplay[index].innerHTML = formatBytes(data.size);
-          dateDisplay[index].innerHTML = data.updated.slice(0,10);
-     }).catch((error) => console.error(error));
-}
-function getWallpaperInfo(f,i){const fileInfoRef = ref(storage, f);getMetadata(fileInfoRef).then(data => {const wallpaperSizeDisplay = document.querySelectorAll("#fileSize");wallpaperSizeDisplay[i].innerHTML = formatBytes(data.size);}).catch((error) => console.error(error));}
 function handleSubmit(e, i){
      e.preventDefault();
      const ids = ["download-option1", "download-option2", "download-option3", "download-option4", "download-option5"];
